hotfix : edit Integer to Float in calc peer review

+++ src/main/java/com/prismworks/prism/domain/prism/dto/PrismDataDto.java
@@ -6,6 +6,6 @@ import java.util.Map;
 
 @Data
 public class PrismDataDto {
-    private Map<String, Integer> prismData;
+    private Map<String, Float> prismData;
     private RadialDataDto radialData;
 }
